weee

+++ Kobold2D-2.0.3/ClusterFear/Projectfiles/SGFoeCluster.h
@@ -51,7 +51,7 @@
     int maxHealth;
     int maxCritters;
     int damage;
-    int moveSpeed;
+    float moveSpeed;
 }
